net: dev: implement the device backlog processor
Implement the backlog processor for the network devices. The backlog is
a linked list containing all packets that are going into or coming out
of the network stack.

The only responsibility that network device drivers bear, is the
allocation of a network buffer and then taking data of the PHY-layer
while copying it into the respective layer buffer. Packets are then to
be enqueued on the backlog of the network device.

The backlog processor will then take packets of the backlog, one by one,
and push RX packets into the `struct netdev::rx()' handler. For ethernet
packets this is `ethernet_input()'. Out going packets are pushed to the
datalink layer by `struct netdev::tx()' and then to the PHY-layer by
`struct netdev::write()'. The network stack shall never free dropped
packets, instead they shall set the `NBUF_DROPPED' flag. The backlog
processor will then free the memory related to this network buffer.
Packets arriving in the application layer will be decoupled from the
backlog through the `NBUF_ARRIVED' flag. The backlog processor will not
free the memory of arrived packets.

Signed-off-by: Michel Megens <dev@bietje.net>