net: dev: increase rx stats only if a packet arrived
The RX stats track the number of bytes that have been succesfully
delivered by the E/STACK network stack. This means they reached their
end-point destination (wether that is application layer, or a network
layer protocol, such as ARP, doesn't matter). Packets that cause an
error will *NOT* be considered when incrementing the RX stats.

Signed-off-by: Michel Megens <dev@bietje.net>