net: dev: clear the RX flag when processing starts
Clear the RX flag on a netbuf after testing it. This way, the buffer can
be reused by the network stack as a TX reply packet without having to
manually twiddle with the packet buffer flags. A function to test and
clear packet buffer flags has been implemented: `netbuf_test_and_clear_flag()'.

Signed-off-by: Michel Megens <dev@bietje.net>