net: dev: run the core processor periodically
The core processor task is ran atleast every `CONFIG_POLL_TMO'
milliseconds. This allows the core to run destination resolving on links
that see barely any traffic. The core processor task still sleeps on an
event and can therefore be woken up more often than `CONFIG_POLL_TMO' by
incoming or outgoing traffic. The value of `CONFIG_POLL_TMO' can be
configured in CMake.

The function `estack_event_wait()' now takes an extra argument that
indicates the duration of the timeout in milliseconds. The timeout can
be set to `FOREVER' to indicate that the event should never timeout. The
return value of `estack_event_wait()' indicates whether the function
received an event or a timeout. Its return codes are:

	* EOK if an event was received
	* ETMO if a timeout has occurred

Signed-off-by: Michel Megens <dev@bietje.net>