net: dev: implement a flag to retain tx packets
Implement the `NBUF_TX_KEEP' flag. Memory allocated to a packet buffer
with this flag set will not be free'd by `netdev_process_backlog()'.
This way protocols that need to maintain their own backlog (e.g. TCP)
can take full control over the memory flow of packet buffers.

Please note, that whenever a packet is transmitted with this flag, that
the networking core will forget about this packet. There is no way this
packet might get released in the future, the protocol layer that had set
this flag is now responsible for the memory associated with this buffer.

Signed-off-by: Michel Megens <dev@bietje.net>