net: core: rewrite the backlog processor
Rewrite the backlog processor in order to process packets more
efficiently. The core processor will test and clear the `NBUF_RX' flag,
and if it was set, the `NBUF_WAS_RX' flag will be set. Flags that are RX
or TX only are also only tested if the packet buffer is an incoming or
outgoing packet respectively.

Before a packet buffer is handed over to the device driver, the packet
buffer is linearized. This means that all network data is stored within
the *same* memory region. Reused RX packets generally don't need this,
as incoming packets are already linear in most cases.

Signed-off-by: Michel Megens <dev@bietje.net>